+++ tests/integration/conversations/v1/conversation/test_participant.py
@@ -36,7 +36,7 @@ class ParticipantTestCase(IntegrationTestCase):
                 "account_sid": "ACaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
                 "conversation_sid": "CHaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
                 "sid": "MBaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
-                "identity": "null",
+                "identity": null,
                 "attributes": "{ \\"role\\": \\"driver\\" }",
                 "messaging_binding": {
                     "type": "sms",
@@ -142,13 +142,13 @@ class ParticipantTestCase(IntegrationTestCase):
                 "account_sid": "ACaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
                 "conversation_sid": "CHaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
                 "sid": "MBaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
-                "identity": "null",
+                "identity": null,
                 "attributes": "{}",
                 "messaging_binding": {
                     "type": "sms",
                     "address": "+15017122661"
                 },
-                "role_sid": "null",
+                "role_sid": null,
                 "date_created": "2020-07-01T22:18:37Z",
                 "date_updated": "2020-07-01T22:18:37Z",
                 "url": "https://conversations.twilio.com/v1/Conversations/CHaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Participants/MBaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa"
@@ -183,7 +183,7 @@ class ParticipantTestCase(IntegrationTestCase):
                 "account_sid": "ACaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
                 "conversation_sid": "CHaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
                 "sid": "MBaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
-                "identity": "null",
+                "identity": null,
                 "attributes": "{ \\"role\\": \\"driver\\" }",
                 "messaging_binding": {
                     "type": "sms",
@@ -275,7 +275,7 @@ class ParticipantTestCase(IntegrationTestCase):
                 "account_sid": "ACaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
                 "conversation_sid": "CHaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
                 "sid": "MBaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
-                "identity": "null",
+                "identity": null,
                 "attributes": "{ \\"role\\": \\"driver\\" }",
                 "messaging_binding": {
                     "type": "sms",
@@ -326,7 +326,7 @@ class ParticipantTestCase(IntegrationTestCase):
                         "account_sid": "ACaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
                         "conversation_sid": "CHaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
                         "sid": "MBaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
-                        "identity": "null",
+                        "identity": null,
                         "attributes": "{ \\"role\\": \\"driver\\" }",
                         "messaging_binding": {
                             "type": "sms",